Add files via upload
These graphs depict the data that was analyzed.
First, two bar plots show the number of people in each group depending on the amount paid. We can see the trend that if the ticket fare grows, the chances to survive will increase. 
The last box plot was maid to show the average age of each fare category and was used to assign new values to NA fields in Age column.